requirements.txt
# 核心依赖
rumps==0.4.0
psutil==5.9.5

# 图像处理（用于图标生成）
Pillow>=9.0.0

# 网络请求
requests>=2.26.0

# 股票数据获取
akshare>=1.8.0

